+++ src/functions/Function.java
@@ -7,10 +7,10 @@ import functions.commutative.Sum;
 import functions.commutative.Product;
 import functions.special.Constant;
 import functions.special.Variable;
+import functions.unitary.Ln;
 import functions.unitary.UnitaryFunction;
 import org.jetbrains.annotations.NotNull;
 import tools.MiscTools;
-import tools.integral.Integral;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -22,7 +22,7 @@ public abstract class Function implements Evaluable, Differentiable, Simplifiabl
 	 */
 	@SuppressWarnings("ClassReferencesSubclass")
 	public static final Class<?>[] sortOrder = {
-			Integral.class,
+			Ln.Integral.class,
 			Constant.class,
 			Variable.class,
 			Product.class,
