+++ tests/IntegralTest.java
@@ -1,10 +1,10 @@
 import functions.Function;
 import functions.commutative.Sum;
+import functions.unitary.Ln;
 import org.junit.jupiter.api.Test;
 import parsing.Parser;
 import tools.exceptions.IntegrationFailedException;
 import tools.helperclasses.Pair;
-import tools.integral.Integral;
 import tools.integral.IntegralsTools;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -14,8 +14,8 @@ public class IntegralTest {
 
     @Test
     void splitAdds() {
-        Integral test1 = new Integral(Parser.parse("x+e^x"), 'x');
-        Function test2 = new Sum(new Integral(Parser.parse("x"), 'x'), new Integral(Parser.parse("e^x"), 'x'));
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("x+e^x"), 'x');
+        Function test2 = new Sum(new Ln.Integral(Parser.parse("x"), 'x'), new Ln.Integral(Parser.parse("e^x"), 'x'));
         assertEquals(test1.integrate() , test2);
     }
 
@@ -29,76 +29,76 @@ public class IntegralTest {
 
     @Test
     void simpleExponentIntegral() {
-        Integral test1 = new Integral(Parser.parse("e^x"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("e^x"), 'x');
         Function test2 = Parser.parse("e^x");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleExponentIntegralWithConstant() {
-        Integral test1 = new Integral(Parser.parse("2e^x"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("2e^x"), 'x');
         Function test2 = Parser.parse("2e^x");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleExponentIntegralWithLinerTermInExponent() {
-        Integral test1 = new Integral(Parser.parse("2e^(2x-3)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("2e^(2x-3)"), 'x');
         Function test2 = Parser.parse("e^(2x-3)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simplePowers() {
-        Integral test1 = new Integral(Parser.parse("x^2"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("x^2"), 'x');
         Function test2 = Parser.parse("1/3 * x^3");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleTrig() {
-        Integral test1 = new Integral(Parser.parse("cos(x)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("cos(x)"), 'x');
         Function test2 = Parser.parse("sin(x)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleTrigWithConstants() {
-        Integral test1 = new Integral(Parser.parse("4*cos(2x+3)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("4*cos(2x+3)"), 'x');
         Function test2 = Parser.parse("2*sin(2x+3)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleReciprocal() {
-        Integral test1 = new Integral(Parser.parse("e^x/(1+e^x)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("e^x/(1+e^x)"), 'x');
         Function test2 = Parser.parse("ln(1+e^x)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleSquareRoot() {
-        Integral test1 = new Integral(Parser.parse("x*sqrt(1+x^2)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("x*sqrt(1+x^2)"), 'x');
         Function test2 = Parser.parse("1/3 * (1+x^2)^(3/2)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleExpUSub() {
-        Integral test1 = new Integral(Parser.parse("x*e^(x^2)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("x*e^(x^2)"), 'x');
         Function test2 = Parser.parse("1/2*e^(x^2)");
         assertEquals(test2, test1.integrate());
     }
 
     @Test
     void simpleExpUSubThatNoWork() {
-        Integral test1 = new Integral(Parser.parse("2x*sin(x)*e^(x^2)"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("2x*sin(x)*e^(x^2)"), 'x');
         assertThrows(IntegrationFailedException.class, test1::integrate);
     }
 
     @Test
     void simpleOPIsOne() {
-        Function test1 = new Integral(Parser.parse("cos(x)*sin(x)"), 'x').integrate();
+        Function test1 = new Ln.Integral(Parser.parse("cos(x)*sin(x)"), 'x').integrate();
         Function test2 = Parser.parse("1/2*(sin(x))^2");
         Function test3 = Parser.parse("-1/2*(cos(x))^2");
         assertTrue(test1.equals(test2) || test1.equals(test3));
@@ -106,7 +106,7 @@ public class IntegralTest {
 
     @Test
     void complexUSub() {
-        Integral test1 = new Integral(Parser.parse("(cos(e^x))^2*sin(e^x)*e^x"), 'x');
+        Ln.Integral test1 = new Ln.Integral(Parser.parse("(cos(e^x))^2*sin(e^x)*e^x"), 'x');
         Function test2 = Parser.parse("-1/3*(cos(e^x))^3");
         assertEquals(test2, test1.integrate());
     }
